Move SpanData to main API and add Resource.
