+++ api/src/main/java/openconsensus/trace/Tracer.java
@@ -20,7 +20,6 @@ import com.google.errorprone.annotations.MustBeClosed;
 import java.util.concurrent.Callable;
 import javax.annotation.Nullable;
 import openconsensus.context.Scope;
-import openconsensus.trace.data.SpanData;
 import openconsensus.trace.propagation.BinaryFormat;
 import openconsensus.trace.propagation.TextFormat;
